+++ src/main/java/org/folio/marc4ld/service/marc2ld/bib/mapper/additional/GeographicCoverageMapper.java
@@ -1,5 +1,6 @@
 package org.folio.marc4ld.service.marc2ld.bib.mapper.additional;
 
+import static org.folio.ld.dictionary.PredicateDictionary.GEOGRAPHIC_COVERAGE;
 import static org.folio.ld.dictionary.PropertyDictionary.GEOGRAPHIC_AREA_CODE;
 import static org.folio.ld.dictionary.PropertyDictionary.NAME;
 import static org.folio.marc4ld.util.Constants.Dictionary.GEOGRAPHIC_CODE_TO_NAME;
@@ -7,7 +8,7 @@ import static org.folio.marc4ld.util.Constants.TAG_043;
 
 import java.util.List;
 import java.util.Set;
-import org.folio.ld.dictionary.PredicateDictionary;
+import org.folio.marc4ld.configuration.property.Marc4LdRules;
 import org.folio.marc4ld.service.dictionary.DictionaryProcessor;
 import org.folio.marc4ld.service.marc2ld.mapper.MapperHelper;
 import org.springframework.stereotype.Component;
@@ -29,8 +30,8 @@ public class GeographicCoverageMapper extends DictionaryBasedMarc2LdMapper {
   }
 
   @Override
-  public boolean canMap(PredicateDictionary predicate) {
-    return predicate == PredicateDictionary.GEOGRAPHIC_COVERAGE;
+  public boolean canMap(Marc4LdRules.FieldRule fieldRule) {
+    return GEOGRAPHIC_COVERAGE.name().equals(fieldRule.getPredicate());
   }
 
   @Override